Refactor multview toolbar components: remove unused files and move liveChannel and liveStreamInfo
